+++ CHANGES.rst
@@ -10,6 +10,8 @@ Changelog
 
 - Add an exception for the HTTP status code 418
 
+- Don't override the `content-type` header if set.
+
 
 4.1 (2018-10-05)
 ----------------